move spawn functions refactor components
