+++ assets/data/enemies.ron
@@ -13,13 +13,13 @@
             defense_damage: 30,
             loot_probs: [
                 (None, 0.5),
-                (Some(HealthWrench), 0.03),
-                (Some(DefenseWrench), 0.02),
-                (Some(Money1), 0.40),
-                (Some(Money5), 0.02),
-                (Some(Armor), 0.03),
+                (Some(Consumable(HealthWrench)), 0.03),
+                (Some(Consumable(DefenseWrench)), 0.02),
+                (Some(Consumable(Money1)), 0.40),
+                (Some(Consumable(Money5)), 0.02),
+                (Some(Consumable(Armor)), 0.03),
             ],
-            entity_type: Drone,
+            entity_type: Enemy(Drone),
         ),
         health_component: (
             max_value: 60,
@@ -78,13 +78,13 @@
             defense_damage: 50,
             loot_probs: [
                 (None, 0.5),
-                (Some(HealthWrench), 0.03),
-                (Some(DefenseWrench), 0.02),
-                (Some(Money1), 0.40),
-                (Some(Money5), 0.02),
-                (Some(Armor), 0.03),
+                (Some(Consumable(HealthWrench)), 0.03),
+                (Some(Consumable(DefenseWrench)), 0.02),
+                (Some(Consumable(Money1)), 0.40),
+                (Some(Consumable(Money5)), 0.02),
+                (Some(Consumable(Armor)), 0.03),
             ],
-            entity_type: Pawn,
+            entity_type: Enemy(Pawn),
         ),
         health_component: (
             max_value: 110,
@@ -126,7 +126,7 @@
             loot_probs: [
                 (None, 1.0),
             ],
-            entity_type: Hauler,
+            entity_type: Enemy(Hauler),
         ),
         health_component: (
             max_value: 220,
@@ -186,13 +186,13 @@
             max_speed: 25,
             loot_probs: [
                 (None, 0.5),
-                (Some(HealthWrench), 0.03),
-                (Some(DefenseWrench), 0.02),
-                (Some(Money1), 0.40),
-                (Some(Money5), 0.02),
-                (Some(Armor), 0.03),
+                (Some(Consumable(HealthWrench)), 0.03),
+                (Some(Consumable(DefenseWrench)), 0.02),
+                (Some(Consumable(Money1)), 0.40),
+                (Some(Consumable(Money5)), 0.02),
+                (Some(Consumable(Armor)), 0.03),
             ],
-            entity_type: Strafer,
+            entity_type: Enemy(Strafer),
         ),
         health_component: (
             max_value: 75,
@@ -234,13 +234,13 @@
             max_speed: 10,
             loot_probs: [
                 (None, 0.5),
-                (Some(HealthWrench), 0.03),
-                (Some(DefenseWrench), 0.02),
-                (Some(Money1), 0.40),
-                (Some(Money5), 0.02),
-                (Some(Armor), 0.03),
+                (Some(Consumable(HealthWrench)), 0.03),
+                (Some(Consumable(DefenseWrench)), 0.02),
+                (Some(Consumable(Money1)), 0.40),
+                (Some(Consumable(Money5)), 0.02),
+                (Some(Consumable(Armor)), 0.03),
             ],
-            entity_type: MissileLauncher,
+            entity_type: Enemy(MissileLauncher),
         ),
         health_component: (
             max_value: 300,
@@ -267,7 +267,7 @@
             bottom_offset: Some(-20.0),
         ),
         auto_child_enemy_spawner_component: Some((
-            child_entity_type: Missile,
+            child_entity_type: Enemy(Missile),
             offset: [0.0, -20.0],
             period: 3.5,
             timer: 3.5,
@@ -290,7 +290,7 @@
             loot_probs: [
                 (None, 1.0),
             ],
-            entity_type: Missile,
+            entity_type: Enemy(Missile),
         ),
         health_component: (
             max_value: 50,
@@ -331,13 +331,13 @@
             defense_damage: 0,
             loot_probs: [
                 (None, 0.0),
-                (Some(HealthWrench), 0.04),
-                (Some(DefenseWrench), 0.02),
-                (Some(Money1), 0.40),
-                (Some(Money5), 0.5),
-                (Some(Armor), 0.03),
+                (Some(Consumable(HealthWrench)), 0.04),
+                (Some(Consumable(DefenseWrench)), 0.02),
+                (Some(Consumable(Money1)), 0.40),
+                (Some(Consumable(Money5)), 0.5),
+                (Some(Consumable(Armor)), 0.03),
             ],
-            entity_type: RepeaterBody,
+            entity_type: Enemy(RepeaterBody),
         ),
         health_component: (
             max_value: 10000,
@@ -378,13 +378,13 @@
             defense_damage: 0,
             loot_probs: [
                 (None, 0.0),
-                (Some(HealthWrench), 0.04),
-                (Some(DefenseWrench), 0.02),
-                (Some(Money1), 0.40),
-                (Some(Money5), 0.5),
-                (Some(Armor), 0.03),
+                (Some(Consumable(HealthWrench)), 0.04),
+                (Some(Consumable(DefenseWrench)), 0.02),
+                (Some(Consumable(Money1)), 0.40),
+                (Some(Consumable(Money5)), 0.5),
+                (Some(Consumable(Armor)), 0.03),
             ],
-            entity_type: RepeaterHead,
+            entity_type: Enemy(RepeaterHead),
         ),
         health_component: (
             max_value: 2000,
@@ -425,13 +425,13 @@
             defense_damage: 0,
             loot_probs: [
                 (None, 0.0),
-                (Some(HealthWrench), 0.04),
-                (Some(DefenseWrench), 0.02),
-                (Some(Money1), 0.40),
-                (Some(Money5), 0.5),
-                (Some(Armor), 0.03),
+                (Some(Consumable(HealthWrench)), 0.04),
+                (Some(Consumable(DefenseWrench)), 0.02),
+                (Some(Consumable(Money1)), 0.40),
+                (Some(Consumable(Money5)), 0.5),
+                (Some(Consumable(Armor)), 0.03)
             ],
-            entity_type: RepeaterRightShoulder,
+            entity_type: Enemy(RepeaterRightShoulder),
         ),
         health_component: (
             max_value: 2000,
@@ -474,13 +474,13 @@
             defense_damage: 0,
             loot_probs: [
                 (None, 0.0),
-                (Some(HealthWrench), 0.04),
-                (Some(DefenseWrench), 0.02),
-                (Some(Money1), 0.40),
-                (Some(Money5), 0.5),
-                (Some(Armor), 0.03),
+                (Some(Consumable(HealthWrench)), 0.04),
+                (Some(Consumable(DefenseWrench)), 0.02),
+                (Some(Consumable(Money1)), 0.40),
+                (Some(Consumable(Money5)), 0.5),
+                (Some(Consumable(Armor)), 0.03)
             ],
-            entity_type: RepeaterLeftShoulder,
+            entity_type: Enemy(RepeaterLeftShoulder),
         ),
         health_component: (
             max_value: 2000,
@@ -524,13 +524,13 @@
             current_velocity_y: -30,
             loot_probs: [
                 (None, 0.0),
-                (Some(HealthWrench), 0.04),
-                (Some(DefenseWrench), 0.02),
-                (Some(Money1), 0.40),
-                (Some(Money5), 0.5),
-                (Some(Armor), 0.03),
+                (Some(Consumable(HealthWrench)), 0.04),
+                (Some(Consumable(DefenseWrench)), 0.02),
+                (Some(Consumable(Money1)), 0.40),
+                (Some(Consumable(Money5)), 0.5),
+                (Some(Consumable(Armor)), 0.03),
             ],
-            entity_type: RepeaterRightArm,
+            entity_type: Enemy(RepeaterRightArm),
         ),
         health_component: (
             max_value: 2000,
