+++ src/components/enemy.rs
@@ -12,15 +12,12 @@ pub struct EnemyComponent {
     pub defense_damage: f32,
     #[serde(default = "des_collision_damage")]
     pub collision_damage: f32,
-    #[serde(default = "des_poison")]
-    pub poison: f32,
     #[serde(default = "des_allied")]
     pub allied: bool,
     pub loot_probs: SpawnProbabilities,
     pub entity_type: EntityType,
     #[serde(default = "des_explosion_sprite_idx")]
     pub explosion_sprite_idx: usize,
-    pub target_position: Option<Vector2<f32>>,
 }
 
 fn des_explosion_sprite_idx() -> usize {
@@ -28,9 +25,6 @@ fn des_explosion_sprite_idx() -> usize {
 }
 fn des_collision_damage() -> f32 {
     30.0
-}
-fn des_poison() -> f32 {
-    0.0
 }
 fn des_allied() -> bool {
     false
