--- src/entities/missile.rs
@@ -1,32 +1,0 @@
-use amethyst::{
-    core::transform::Transform,
-    ecs::prelude::{Builder, Entities, LazyUpdate, ReadExpect},
-    renderer::{SpriteRender, Transparent},
-};
-
-use crate::components::{BlastComponent, Hitbox2DComponent, Motion2DComponent};
-
-// spaces and creates blast entities
-pub fn spawn_missiles(
-    blast_sprite_render: SpriteRender,
-    blast_component: BlastComponent,
-    blast_hitbox: Hitbox2DComponent,
-    blast_motion2d: Motion2DComponent,
-    mut blast_transform: Transform,
-    entities: &Entities,
-    lazy_update: &ReadExpect<LazyUpdate>,
-) {
-    for _ in 0..blast_count {
-        lazy_update
-            .create_entity(entities)
-            .with(blast_component.clone())
-            .with(blast_hitbox.clone())
-            .with(blast_motion2d.clone())
-            .with(blast_sprite_render.clone())
-            .with(blast_transform.clone())
-            .with(Transparent)
-            .build();
-
-        blast_transform.prepend_translation_x(blast_spacing);
-    }
-}
