+++ src/entities/mod.rs
@@ -6,13 +6,13 @@ pub mod blast;
 pub mod boss;
 pub mod consumable;
 pub mod defense;
-pub mod enemy;
 pub mod enemy_spawner;
 pub mod explosion;
 pub mod items;
 pub mod planet;
 pub mod side_panels;
 pub mod spaceship;
+pub mod spawn;
 pub mod status_bar;
 pub mod status_unit;
 pub mod store;
@@ -22,15 +22,14 @@ pub use self::{
     barriers::initialize_arena_barriers,
     blast::spawn_blasts,
     boss::spawn_repeater,
-    consumable::{spawn_consumable, spawn_random_consumable},
+    consumable::spawn_random_consumable,
     defense::initialize_defense,
-    enemy::spawn_enemy,
     enemy_spawner::initialize_enemy_spawner,
     explosion::{spawn_blast_explosion, spawn_explosion},
-    items::spawn_item,
     planet::initialize_planet,
     side_panels::initialize_side_panels,
     spaceship::initialize_spaceship,
+    spawn::{spawn_consumable, spawn_enemy, spawn_item},
     status_bar::initialize_status_bars,
     status_unit::spawn_status_unit,
     store::initialize_store,
@@ -38,7 +37,13 @@ pub use self::{
 
 #[derive(Clone, Serialize, Deserialize, Debug, Hash, PartialEq, Eq)]
 pub enum EntityType {
-    // Enemies
+    Enemy(EnemyType),
+    Consumable(ConsumableType),
+    Item(ItemType),
+}
+
+#[derive(Clone, Serialize, Deserialize, Debug, Hash, PartialEq, Eq)]
+pub enum EnemyType {
     Pawn,
     Drone,
     Strafer,
@@ -51,13 +56,19 @@ pub enum EntityType {
     RepeaterRightShoulder,
     RepeaterLeftArm,
     RepeaterRightArm,
-    // Consumables
+}
+
+#[derive(Clone, Serialize, Deserialize, Debug, Hash, PartialEq, Eq)]
+pub enum ConsumableType {
     DefenseWrench,
     Money1,
     Money5,
     HealthWrench,
     Armor,
-    // Items
+}
+
+#[derive(Clone, Serialize, Deserialize, Debug, Hash, PartialEq, Eq)]
+pub enum ItemType {
     SteelBarrel,
     PlasmaBlasts,
     HazardousReactor,
